@@ -39,22 +39,24 @@
   "peerDependencies": {
     "@agm/core": "^1.0.0-beta.7",
     "dap-query-js": "github:flowmatters/dap-query-js",
-    "ngx-pipes": "1.6.5",
+    "ngx-pipes": "^2.7.5",
     "plotly.js": "git+https://github.com/arlowhite/plotly.js#no-polyfill",
     "proj4": "^2.3.17"
   },
   "devDependencies": {
     "@agm/core": "^1.0.0-beta.7",
-    "@angular/animations": "^8.2.10",
-    "@angular/common": "^8.2.10",
-    "@angular/compiler": "^8.2.10",
-    "@angular/compiler-cli": "^8.2.10",
-    "@angular/core": "^8.2.10",
-    "@angular/forms": "^8.2.10",
-    "@angular/platform-browser": "^8.2.10",
-    "@angular/platform-browser-dynamic": "^8.2.10",
-    "@angular/platform-server": "^8.2.10",
-    "@angular/router": "^8.2.10",
+    "@angular-devkit/schematics": "^11.0.2",
+    "@angular/animations": "^11.0.2",
+    "@angular/cli": "^11.0.2",
+    "@angular/common": "^11.0.2",
+    "@angular/compiler": "^11.0.2",
+    "@angular/compiler-cli": "^11.0.2",
+    "@angular/core": "^11.0.2",
+    "@angular/forms": "^11.0.2",
+    "@angular/platform-browser": "^11.0.2",
+    "@angular/platform-browser-dynamic": "^11.0.2",
+    "@angular/platform-server": "^11.0.2",
+    "@angular/router": "^11.0.2",
     "@ng-bootstrap/ng-bootstrap": "^4.0.0",
     "@types/es6-shim": "^0.31.32",
     "@types/geojson": "^1.0.3",
@@ -70,15 +72,15 @@
     "map-wald": "github:ANU-WALD/map-wald",
     "merge": "^1.2.1",
     "ng-packagr": "^5.5.1",
-    "ngx-pipes": "1.6.5",
+    "ngx-pipes": "^2.7.5",
     "plotly.js": "git+https://github.com/arlowhite/plotly.js#no-polyfill",
     "proj4": "^2.3.17",
     "rxjs": "^6.5.3",
     "store": "^2.0.12",
-    "tslint": "~5.11.0",
-    "typescript": "~3.5.3",
+    "tslint": "~6.1.0",
+    "typescript": "~4.0.5",
     "watch": "latest",
-    "zone.js": "0.9.1"
+    "zone.js": "0.10.3"
   },
   "engines": {
     "node": ">=0.8.0"
